Refactor GeoNode backup admin command
- Add command for GeoNode backup for QGIS Server backend
- Refactor default management command and Admin command to execute
  backup/restore command based on OGC backend being used
