+++ geonode/qgis_server/management/commands/settings.ini
@@ -1,0 +1,28 @@
+[database]
+pgdump = pg_dump
+pgrestore = pg_restore
+
+[qgis_server]
+datadir = geonode/qgis_layer
+
+[fixtures]
+# NOTE: Order is important
+# 2.4
+# apps   = people,account,avatar.avatar,base.backup,base.license,base.topiccategory,base.region,base.restrictioncodetype,base.spatialrepresentationtype,base.resourcebase,base.contactrole,base.link,guardian.userobjectpermission,guardian.groupobjectpermission,layers.uploadsession,layers.style,layers.layer,layers.attribute,layers.layerfile,maps.map,maps.maplayer,maps.mapsnapshot,documents.document,taggit
+# dumps  = people,accounts,avatars,backups,licenses,topiccategories,regions,spatialrepresentationtypes,restrictioncodetypes,resourcebases,contactroles,links,useropermissions,groupopermissions,uploadsessions,styles,layers,attributes,layerfiles,maps,maplayers,mapsnapshots,documents,tags
+
+# 2.6+
+apps  = contenttypes,auth,people,groups,account,actstream,admin,agon_ratings,announcements,avatar,base,dialogos,documents,guardian,invitations,layers,maps,qgis_server,oauth2_provider,pinax_notifications,services,sessions,sites,taggit,upload,user_messages
+dumps = contenttypes,auth,people,groups,account,actstream,admin,agon_ratings,announcements,avatar,base,dialogos,documents,guardian,invitations,layers,maps,qgis_server,oauth2_provider,pinax_notifications,services,sessions,sites,taggit,upload,user_messages
+
+# Migrate from GN 2.0 to GN 2.4
+# migrations = base.resourcebase,layers.layer,layers.attribute,maps.map,maps.maplayer
+# manglers   = gn20_to_24.ResourceBaseMangler,gn20_to_24.LayerMangler,gn20_to_24.LayerAttributesMangler,gn20_to_24.MapMangler,gn20_to_24.MapLayersMangler
+
+# Migrate from GN 2.4 to GN 2.4
+# migrations = base.resourcebase,layers.layer,layers.attribute,maps.map,maps.maplayer
+# manglers   = gn24_to_24.ResourceBaseMangler,gn24_to_24.LayerMangler,gn24_to_24.LayerAttributesMangler,gn24_to_24.DefaultMangler,gn24_to_24.MapLayersMangler
+
+# Migrate from GN 2.6+
+migrations = base.resourcebase,layers.layer,layers.attribute,maps.map,maps.maplayer
+manglers   = gn24_to_24.ResourceBaseMangler,gn24_to_24.LayerMangler,gn24_to_24.LayerAttributesMangler,gn24_to_24.DefaultMangler,gn24_to_24.MapLayersMangler
